+++ src/Command/SendDailyNotificationsCommand/TelegramConfigurationReminderGetter.php
@@ -22,8 +22,11 @@ class TelegramConfigurationReminderGetter implements NotificationGetterInterface
         foreach ($telegram_notifications_state as $user_id => $state) {
             $reminder_id = $state['reminder_id'] ?? false;
             $needs_reminder = $state['needs_reminder'] ?? false;
-            if ($needs_reminder && !$reminder_id) {
-                $user = $user_repo->findOneBy(['id' => $user_id]);
+            $user = $user_repo->findOneBy(['id' => $user_id]);
+            if (!$user) {
+                $this->log()->warning("No user (ID:{$user_id}) for telegram notification");
+            }
+            if ($needs_reminder && !$reminder_id && $user) {
                 $this->log()->info("Generating telegram configuration reminder subscription for '{$user}'...");
                 $subscription = new NotificationSubscription();
                 $subscription->setUser($user);
@@ -34,10 +37,11 @@ class TelegramConfigurationReminderGetter implements NotificationGetterInterface
                 $this->entityManager()->persist($subscription);
             }
             if ($reminder_id && !$needs_reminder) {
-                $user = $user_repo->findOneBy(['id' => $user_id]);
                 $this->log()->info("Removing telegram configuration reminder subscription ({$reminder_id}) for '{$user}'...");
                 $subscription = $notification_subscription_repo->findOneBy(['id' => $reminder_id]);
-                $this->entityManager()->remove($subscription);
+                if ($subscription) {
+                    $this->entityManager()->remove($subscription);
+                }
             }
         }
         $this->entityManager()->flush();
@@ -55,7 +59,9 @@ class TelegramConfigurationReminderGetter implements NotificationGetterInterface
         foreach ($telegram_notification_subscriptions as $subscription) {
             $user_id = $subscription->getUser()->getId() ?: 0;
             $user_state = $telegram_notifications_state[$user_id] ?? [];
-            $user_state['reminder_id'] = $subscription->getId();
+            $subscription_id = $subscription->getId();
+            $this->generalUtils()->checkNotNull($subscription_id, "No subscription ID");
+            $user_state['reminder_id'] = $subscription_id;
             $telegram_notifications_state[$user_id] = $user_state;
         }
